requirements.txt
newspaper3k
transformers
torch
matplot
pandas 
nlp
numpy
seaborn